[fix]filterコンポーネント内を再利用するよう変更
